Move calc_normalized_p to lib/failure_detection.py
Consolidate the function with its primary usage location.
